feat: optional storyblok title and description fields that overwrite product content
