+++ instant.config.json
@@ -6,6 +6,9 @@
 		},
 		"SocialProof": {
 			"id": "4adf1b33-0714-4c17-a250-7d706991cade"
+		},
+		"ProductCta": {
+			"id": "1b9f3baa-a218-4456-8aac-adfa1aae2e89"
 		}
 	}
 }
