+++ src/lib/shopify/index.ts
@@ -1,1 +1,2 @@
-export * from './products';
+export * from "./products";
+export * from "./product";
