+++ src/lib/shopify/product.ts
@@ -1,0 +1,32 @@
+import { gql } from "@instantcommerce/sdk";
+
+export interface Product {
+  title: string;
+  description: string;
+  variants: {
+    edges: any[];
+  };
+}
+
+export interface ProductResponse {
+  product: Product;
+}
+
+export const productQuery = gql`
+  query getProductById($id: ID!) {
+    product(id: $id) {
+      title
+      description
+      ... on Product {
+        variants(first: 100) {
+          edges {
+            node {
+              id
+              price
+            }
+          }
+        }
+      }
+    }
+  }
+`;
